+++ crates/stackable-operator/src/product_config_utils.rs
@@ -8,6 +8,9 @@ use tracing::{debug, error, warn};
 
 use crate::role_utils::{CommonConfiguration, Role};
 
+pub const CONFIG_OVERRIDE_FILE_HEADER_KEY: &str = "EXPERIMENTAL_FILE_HEADER";
+pub const CONFIG_OVERRIDE_FILE_FOOTER_KEY: &str = "EXPERIMENTAL_FILE_FOOTER";
+
 type Result<T, E = Error> = std::result::Result<T, E>;
 
 #[derive(Debug, PartialEq, Snafu)]